requirements.txt
fastapi==0.110.0
strawberry-graphql[fastapi]==0.307.1
uvicorn==0.41.0
motor==3.3.2
pymongo==4.6.0
